Rely on flock for QLockFile to solve stale locks and fsync overhead
QLockFile is using a complicated sequence of operations to acquire
locks and a series of heuristics to detect stale locks. In the way these
are implemented for unix, it's possible for stale locks to appear locked
for up to 30 seconds and becomes even more fragile if the system clock
and mtime are not aligned. Blocking to acquire in QLockFile is common
behavior, including by QSettings while writing ini files. This can cause
enormous delays on the UI thread.

Empty lock files are an especially ugly case. They have to be treated as
locked, but the usual process heuristics aren't available, so they get
the worst case of delays. To reduce that possibility, QLockFile calls
f(data)sync after writing. With heavy IO load that can take up to
several seconds, and even in moderate situations it commonly takes tens
to hundreds of milliseconds.

File locking in unix is broken in many ways, but for this purpose on a
modern Linux system, I think either flock() or F_OFD_SETLK is sufficient
for a reliable lock file. The POSIX F_SETLK is not usable.

Criticism of the remaining two is related to behavior over NFS and,
oddly, the relationship between different types of advisory locks. NFS
behavior for either seems to be well defined and acceptable under modern
Linux, and multiple locking APIs are irrelevant: QLockFile is making no
promises about interoperability with other implementations, and the edge
case of mixed versions of Qt could be managed in other ways.

This commit rewrites qlockfile_unix to rely on an exclusive flock() of
the lock file, disregarding the existence of the file or its contents.
This makes stale locks impossible and removes the need for fsync,
solving both of these major cases of unnecessary blocking.

Change-Id: Id6dfb487f60d7a273739be5475ae518dbb25733c