Disable fsync for QSaveFile to avoid IO problems despite risks
For our system, fsync() can cause extreme delays during IO load, on the
order of multiple seconds. Lacking a good solution to that at the system
level right now, making QSaveFile a little bit unsafe is necessary to
solve UI lag/timeout issues.

This creates a small risk of data loss with QSaveFile. If power is lost
or there is a kernel/hardware failure while writing the file that occurs
exactly between flushing the filesystem metadata to disk and flushing
the file data itself, the target may be replaced by a 0 byte file. In
context, this seems like an acceptable tradeoff.

QSaveFile still has its other benefits, including the atomic rename to
update the file.

This patch is not suitable for upstream and should be discarded if fsync
performance under load becomes reasonable.